+++ elasticsearch5/src/main/java/org/vertexium/elasticsearch5/Elasticsearch5ExceptionHandler.java
@@ -1,0 +1,15 @@
+package org.vertexium.elasticsearch5;
+
+import org.vertexium.Graph;
+import org.vertexium.elasticsearch5.utils.FlushObjectQueue;
+
+public interface Elasticsearch5ExceptionHandler {
+    default void handleDocumentMissingException(
+            Graph graph,
+            Elasticsearch5SearchIndex elasticsearch5SearchIndex,
+            FlushObjectQueue.FlushObject flushObject,
+            Exception ex
+    ) throws Exception {
+        throw ex;
+    }
+}