+++ multimodule-test/accumulo-elasticsearch5/src/test/java/org/vertexium/multimodule/AccumuloElasticsearch5Test.java
@@ -5,16 +5,22 @@ import org.apache.accumulo.core.client.AccumuloException;
 import org.apache.accumulo.core.client.AccumuloSecurityException;
 import org.junit.Before;
 import org.junit.ClassRule;
+import org.junit.Test;
 import org.vertexium.Graph;
+import org.vertexium.GraphWithSearchIndex;
+import org.vertexium.Vertex;
 import org.vertexium.VertexiumException;
 import org.vertexium.accumulo.AccumuloGraph;
 import org.vertexium.accumulo.AccumuloGraphConfiguration;
 import org.vertexium.accumulo.AccumuloGraphTestBase;
 import org.vertexium.accumulo.AccumuloResource;
+import org.vertexium.elasticsearch5.Elasticsearch5SearchIndex;
 import org.vertexium.elasticsearch5.ElasticsearchResource;
+import org.vertexium.elasticsearch5.TestElasticsearch5ExceptionHandler;
 import org.vertexium.elasticsearch5.scoring.ElasticsearchFieldValueScoringStrategy;
 import org.vertexium.elasticsearch5.scoring.ElasticsearchHammingDistanceScoringStrategy;
 import org.vertexium.id.SimpleNameSubstitutionStrategy;
+import org.vertexium.mutation.ExistingElementMutation;
 import org.vertexium.scoring.ScoringStrategy;
 
 import java.io.IOException;
@@ -22,9 +28,7 @@ import java.net.URISyntaxException;
 import java.util.HashMap;
 import java.util.Map;
 
-import static org.vertexium.id.SimpleSubstitutionUtils.KEY_IDENTIFIER;
-import static org.vertexium.id.SimpleSubstitutionUtils.SUBSTITUTION_MAP_PREFIX;
-import static org.vertexium.id.SimpleSubstitutionUtils.VALUE_IDENTIFIER;
+import static org.vertexium.id.SimpleSubstitutionUtils.*;
 
 public class AccumuloElasticsearch5Test extends AccumuloGraphTestBase {
     @ClassRule
@@ -39,7 +43,6 @@ public class AccumuloElasticsearch5Test extends AccumuloGraphTestBase {
         put(Joiner.on('.').join(new String[]{SUBSTITUTION_MAP_PREFIX, "3", KEY_IDENTIFIER}), "http://vertexium.org");
         put(Joiner.on('.').join(new String[]{SUBSTITUTION_MAP_PREFIX, "3", VALUE_IDENTIFIER}), "hvo");
     }});
-
 
     @ClassRule
     public static final ElasticsearchResource elasticsearchResource = new ElasticsearchResource(AccumuloElasticsearch5Test.class.getName());
@@ -57,6 +60,10 @@ public class AccumuloElasticsearch5Test extends AccumuloGraphTestBase {
         Map accumuloConfig = accumuloResource.createConfig();
         accumuloConfig.putAll(elasticsearchResource.createConfig());
         return AccumuloGraph.create(new AccumuloGraphConfiguration(accumuloConfig));
+    }
+
+    private Elasticsearch5SearchIndex getSearchIndex() {
+        return (Elasticsearch5SearchIndex) ((GraphWithSearchIndex) graph).getSearchIndex();
     }
 
     @Override
@@ -115,5 +122,22 @@ public class AccumuloElasticsearch5Test extends AccumuloGraphTestBase {
     @Override
     protected ScoringStrategy getFieldValueScoringStrategy(String field) {
         return new ElasticsearchFieldValueScoringStrategy(field);
+    }
+
+    @Test
+    public void testDocumentMissingHandler() throws Exception {
+        TestElasticsearch5ExceptionHandler.authorizations = AUTHORIZATIONS_ALL;
+
+        graph.addVertex("v1", VISIBILITY_A, AUTHORIZATIONS_A);
+        graph.addVertex("v2", VISIBILITY_A, AUTHORIZATIONS_A);
+        graph.flush();
+
+        elasticsearchResource.clearIndices(getSearchIndex());
+
+        Vertex v1 = graph.getVertex("v1", AUTHORIZATIONS_A);
+        ExistingElementMutation<Vertex> m = v1.prepareMutation();
+        m.setProperty("prop1", "value1", VISIBILITY_A);
+        m.save(AUTHORIZATIONS_A);
+        graph.flush();
     }
 }